Class 5

+++ Class 2/Check.java
@@ -3,7 +3,7 @@ class Check
 
 	public static void main(String[] args) 
 	{		
-		int[] b = {1,2,3,4,5};  //0, 1, 2, 3, 4
+		double[] b = {1,2,3,4,5};  //0, 1, 2, 3, 4
 		System.out.print(b[5]);
 	}
